requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
azure-cognitiveservices-vision-customvision>=0.4.0  # sample won't work with previous versions
azure-keyvault>=0.3.6